requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
filelock==3.5.1
idna==3.3
Pillow==9.0.1
platformdirs==2.5.0
PyJWT==2.3.0
pyshorteners==1.0.1
pytz==2021.3
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
twilio==7.9.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
